+++ app/vmui/packages/vmui/src/components/Main/ShortcutKeys/ShortcutKeys.tsx
@@ -1,7 +1,4 @@
-import { FC, useCallback } from "preact/compat";
-import { getAppModeEnable } from "../../../utils/app-mode";
-import Button from "../Button/Button";
-import { KeyboardIcon } from "../Icons";
+import { FC, ReactNode, useCallback } from "preact/compat";
 import Modal from "../Modal/Modal";
 import "./style.scss";
 import Tooltip from "../Tooltip/Tooltip";
@@ -14,8 +11,11 @@ const title = "Shortcut keys";
 const isMac = isMacOs();
 const keyOpenHelp = isMac ? "Cmd + /" : "F1";
 
-const ShortcutKeys: FC<{ showTitle?: boolean }> = ({ showTitle }) => {
-  const appModeEnable = getAppModeEnable();
+type Props = {
+  children?: ReactNode
+}
+
+const ShortcutKeys: FC<Props> = ({ children }) => {
 
   const {
     value: openList,
@@ -35,20 +35,12 @@ const ShortcutKeys: FC<{ showTitle?: boolean }> = ({ showTitle }) => {
 
   return <>
     <Tooltip
-      open={showTitle === true ? false : undefined}
       title={`${title} (${keyOpenHelp})`}
       placement="bottom-center"
     >
-      <Button
-        className={appModeEnable ? "" : "vm-header-button"}
-        variant="contained"
-        color="primary"
-        startIcon={<KeyboardIcon/>}
-        onClick={handleOpen}
-        ariaLabel={title}
-      >
-        {showTitle && title}
-      </Button>
+      <div onClick={handleOpen}>
+        {children}
+      </div>
     </Tooltip>
 
     {openList && (
@@ -62,9 +54,6 @@ const ShortcutKeys: FC<{ showTitle?: boolean }> = ({ showTitle }) => {
               className="vm-shortcuts-section"
               key={section.title}
             >
-              {section.readMore && (
-                <div className="vm-shortcuts-section__read-more">{section.readMore}</div>
-              )}
               <h3 className="vm-shortcuts-section__title">
                 {section.title}
               </h3>