+++ app/vmui/packages/vmui/src/components/Main/ShortcutKeys/constants/keyList.tsx
@@ -1,6 +1,4 @@
 import { isMacOs } from "../../../../utils/detect-device";
-import { VisibilityIcon } from "../../Icons";
-import GraphTips from "../../../Chart/GraphTips/GraphTips";
 
 const ctrlMeta = <code>{isMacOs() ? "Cmd" : "Ctrl"}</code>;
 
@@ -27,10 +25,6 @@ const keyList = [
         description: "Next command from the Query history"
       },
       {
-        keys: <>{ctrlMeta} + <code>click</code> by <VisibilityIcon/></>,
-        description: "Toggle multiple queries"
-      },
-      {
         keys: AUTOCOMPLETE_QUICK_KEY,
         description: "Show quick autocomplete tips"
       }
@@ -38,7 +32,6 @@ const keyList = [
   },
   {
     title: "Graph",
-    readMore: <GraphTips/>,
     list: [
       {
         keys: <>{ctrlMeta} + <code>scroll Up</code> or <code>+</code></>,
@@ -53,12 +46,8 @@ const keyList = [
         description: "Move the graph left/right"
       },
       {
-        keys: <><code>click</code></>,
-        description: "Select the series in the legend"
-      },
-      {
-        keys: <>{ctrlMeta} + <code>click</code></>,
-        description: "Toggle multiple series in the legend"
+        keys: <><code>click</code> on legend item</>,
+        description: "Open the legend item menu"
       }
     ]
   },